Rename yarte_wasm_app to delorean and update
