+++ frameworks/non-keyed/delorean/bundled-dist/index.js
@@ -1,0 +1,1 @@
+!function(e){function n(n){for(var t,o,u=n[0],i=n[1],c=0,_=[];c<u.length;c++)o=u[c],Object.prototype.hasOwnProperty.call(r,o)&&r[o]&&_.push(r[o][0]),r[o]=0;for(t in i)Object.prototype.hasOwnProperty.call(i,t)&&(e[t]=i[t]);for(s&&s(n);_.length;)_.shift()()}var t={},r={0:0};var o={};var u={3:function(){return{"./index_bg.js":{__wbg_preventDefault_93d06688748bfc14:function(e){return t[2].exports.r(e)},__wbindgen_object_drop_ref:function(e){return t[2].exports.D(e)},__wbg_settextContent_9ac5ef9163ad40d0:function(e,n,r){return t[2].exports.x(e,n,r)},__wbg_remove_cc7ef8dcae8cd130:function(e){return t[2].exports.t(e)},__wbindgen_cb_drop:function(e){return t[2].exports.z(e)},__wbg_getstate_58552c3de9b5eb17:function(e){return t[2].exports.j(e)},__wbindgen_object_clone_ref:function(e){return t[2].exports.C(e)},__wbg_instanceof_Window_adf3196bdc02b386:function(e){return t[2].exports.m(e)},__wbg_document_6cc8d0b87c0a99b9:function(e){return t[2].exports.g(e)},__wbg_body_8c888fe47d81765f:function(e){return t[2].exports.c(e)},__wbg_firstElementChild_f5d9e8fe9a5510b7:function(e){return t[2].exports.i(e)},__wbg_nextElementSibling_0fb824df2518e129:function(e){return t[2].exports.o(e)},__wbg_now_4de5b53a19e45567:function(){return t[2].exports.q()},__wbg_createElement_5bdf88a5af9f17c5:function(e,n,r){return t[2].exports.f(e,n,r)},__wbg_appendChild_77215fd672b162c5:function(e,n){return t[2].exports.b(e,n)},__wbg_setAttribute_727bdb9763037624:function(e,n,r,o,u){return t[2].exports.v(e,n,r,o,u)},__wbg_firstChild_da572db3c494324b:function(e){return t[2].exports.h(e)},__wbg_nextSibling_91d4e5504e422251:function(e){return t[2].exports.p(e)},__wbg_addEventListener_9e7b0c3f65ebc0d7:function(e,n,r,o){return t[2].exports.a(e,n,r,o)},__wbg_cloneNode_1db72f2b069d47b9:function(e,n){return t[2].exports.e(e,n)},__wbg_removeEventListener_e118aefce350c930:function(e,n,r,o){return t[2].exports.s(e,n,r,o)},__wbg_setclassName_dfd1832d72e3b25a:function(e,n,r){return t[2].exports.w(e,n,r)},__wbg_self_07b2f89e82ceb76d:function(){return t[2].exports.u()},__wbg_window_ba85d88572adc0dc:function(){return t[2].exports.y()},__wbg_globalThis_b9277fc37e201fe5:function(){return t[2].exports.k()},__wbg_global_e16303fe83e1d57f:function(){return t[2].exports.l()},__wbindgen_is_undefined:function(e){return t[2].exports.B(e)},__wbg_newnoargs_f3b8a801d5d4b079:function(e,n){return t[2].exports.n(e,n)},__wbg_call_8e95613cc6524977:function(e,n){return t[2].exports.d(e,n)},__wbindgen_throw:function(e,n){return t[2].exports.E(e,n)},__wbindgen_closure_wrapper152:function(e,n,r){return t[2].exports.A(e,n,r)}}}}};function i(n){if(t[n])return t[n].exports;var r=t[n]={i:n,l:!1,exports:{}};return e[n].call(r.exports,r,r.exports,i),r.l=!0,r.exports}i.e=function(e){var n=[],t=r[e];if(0!==t)if(t)n.push(t[2]);else{var c=new Promise((function(n,o){t=r[e]=[n,o]}));n.push(t[2]=c);var _,f=document.createElement("script");f.charset="utf-8",f.timeout=120,i.nc&&f.setAttribute("nonce",i.nc),f.src=function(e){return i.p+""+({}[e]||e)+".js"}(e);var s=new Error;_=function(n){f.onerror=f.onload=null,clearTimeout(a);var t=r[e];if(0!==t){if(t){var o=n&&("load"===n.type?"missing":n.type),u=n&&n.target&&n.target.src;s.message="Loading chunk "+e+" failed.\n("+o+": "+u+")",s.name="ChunkLoadError",s.type=o,s.request=u,t[1](s)}r[e]=void 0}};var a=setTimeout((function(){_({type:"timeout",target:f})}),12e4);f.onerror=f.onload=_,document.head.appendChild(f)}return({1:[3]}[e]||[]).forEach((function(e){var t=o[e];if(t)n.push(t);else{var r,c=u[e](),_=fetch(i.p+""+{3:"771be7d21128346d6628"}[e]+".module.wasm");if(c instanceof Promise&&"function"==typeof WebAssembly.compileStreaming)r=Promise.all([WebAssembly.compileStreaming(_),c]).then((function(e){return WebAssembly.instantiate(e[0],e[1])}));else if("function"==typeof WebAssembly.instantiateStreaming)r=WebAssembly.instantiateStreaming(_,c);else{r=_.then((function(e){return e.arrayBuffer()})).then((function(e){return WebAssembly.instantiate(e,c)}))}n.push(o[e]=r.then((function(n){return i.w[e]=(n.instance||n).exports})))}})),Promise.all(n)},i.m=e,i.c=t,i.d=function(e,n,t){i.o(e,n)||Object.defineProperty(e,n,{enumerable:!0,get:t})},i.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},i.t=function(e,n){if(1&n&&(e=i(e)),8&n)return e;if(4&n&&"object"==typeof e&&e&&e.__esModule)return e;var t=Object.create(null);if(i.r(t),Object.defineProperty(t,"default",{enumerable:!0,value:e}),2&n&&"string"!=typeof e)for(var r in e)i.d(t,r,function(n){return e[n]}.bind(null,r));return t},i.n=function(e){var n=e&&e.__esModule?function(){return e.default}:function(){return e};return i.d(n,"a",n),n},i.o=function(e,n){return Object.prototype.hasOwnProperty.call(e,n)},i.p="bundled-dist/",i.oe=function(e){throw console.error(e),e},i.w={};var c=window.webpackJsonp=window.webpackJsonp||[],_=c.push.bind(c);c.push=n,c=c.slice();for(var f=0;f<c.length;f++)n(c[f]);var s=_;i(i.s=0)}([function(e,n,t){t.e(1).then(t.bind(null,1)).catch(console.error)}]);
