--- frameworks/non-keyed/yarte-wasm-app/bundled-dist/1.js
@@ -1,1 +1,0 @@
-(window.webpackJsonp=window.webpackJsonp||[]).push([[1],[,function(n,t,e){"use strict";e.r(t);var r=e(3),u=e(2);e.d(t,"start",(function(){return u.F})),e.d(t,"__wbg_preventDefault_7670dc6ff59bc226",(function(){return u.r})),e.d(t,"__wbindgen_object_drop_ref",(function(){return u.D})),e.d(t,"__wbg_settextContent_d2083d38d155212a",(function(){return u.x})),e.d(t,"__wbg_remove_753943fab80b89c7",(function(){return u.t})),e.d(t,"__wbindgen_cb_drop",(function(){return u.z})),e.d(t,"__wbg_getstate_7e06669f5fd16126",(function(){return u.j})),e.d(t,"__wbindgen_object_clone_ref",(function(){return u.C})),e.d(t,"__wbg_instanceof_Window_e8f84259147dce74",(function(){return u.m})),e.d(t,"__wbg_document_d3b6d86af1c5d199",(function(){return u.g})),e.d(t,"__wbg_body_61c142aa6eae691f",(function(){return u.c})),e.d(t,"__wbg_firstElementChild_f7ee091915e413c4",(function(){return u.i})),e.d(t,"__wbg_nextElementSibling_aab6b4e0e875e3b9",(function(){return u.o})),e.d(t,"__wbg_now_ad54c8b6865a8103",(function(){return u.q})),e.d(t,"__wbg_createElement_d00b8e24838e36e1",(function(){return u.f})),e.d(t,"__wbg_appendChild_8658f795c44d1316",(function(){return u.b})),e.d(t,"__wbg_setAttribute_156f15ecfed9f628",(function(){return u.v})),e.d(t,"__wbg_firstChild_6278272abfd8998c",(function(){return u.h})),e.d(t,"__wbg_nextSibling_051d5a3751133ac7",(function(){return u.p})),e.d(t,"__wbg_addEventListener_116c561435e7160d",(function(){return u.a})),e.d(t,"__wbg_cloneNode_c0331876693c38ac",(function(){return u.e})),e.d(t,"__wbg_removeEventListener_d9ceb7fdf4ca5166",(function(){return u.s})),e.d(t,"__wbg_setclassName_5c90ce56028589a5",(function(){return u.w})),e.d(t,"__wbg_self_179e8c2a5a4c73a3",(function(){return u.u})),e.d(t,"__wbg_window_492cfe63a6e41dfa",(function(){return u.y})),e.d(t,"__wbg_globalThis_8ebfea75c2dd63ee",(function(){return u.k})),e.d(t,"__wbg_global_62ea2619f58bf94d",(function(){return u.l})),e.d(t,"__wbindgen_is_undefined",(function(){return u.B})),e.d(t,"__wbg_newnoargs_e2fdfe2af14a2323",(function(){return u.n})),e.d(t,"__wbg_call_e9f0ce4da840ab94",(function(){return u.d})),e.d(t,"__wbindgen_throw",(function(){return u.E})),e.d(t,"__wbindgen_closure_wrapper161",(function(){return u.A})),r.e()},function(n,t,e){"use strict";(function(n,r){e.d(t,"F",(function(){return y})),e.d(t,"r",(function(){return E})),e.d(t,"D",(function(){return C})),e.d(t,"x",(function(){return j})),e.d(t,"t",(function(){return T})),e.d(t,"z",(function(){return A})),e.d(t,"j",(function(){return D})),e.d(t,"C",(function(){return k})),e.d(t,"m",(function(){return O})),e.d(t,"g",(function(){return P})),e.d(t,"c",(function(){return q})),e.d(t,"i",(function(){return F})),e.d(t,"o",(function(){return L})),e.d(t,"q",(function(){return N})),e.d(t,"f",(function(){return S})),e.d(t,"b",(function(){return B})),e.d(t,"v",(function(){return I})),e.d(t,"h",(function(){return z})),e.d(t,"p",(function(){return J})),e.d(t,"a",(function(){return W})),e.d(t,"e",(function(){return M})),e.d(t,"s",(function(){return U})),e.d(t,"w",(function(){return G})),e.d(t,"u",(function(){return H})),e.d(t,"y",(function(){return K})),e.d(t,"k",(function(){return Q})),e.d(t,"l",(function(){return R})),e.d(t,"B",(function(){return V})),e.d(t,"n",(function(){return X})),e.d(t,"d",(function(){return Y})),e.d(t,"E",(function(){return Z})),e.d(t,"A",(function(){return $}));var u=e(3);const o=new Array(32).fill(void 0);function c(n){return o[n]}o.push(void 0,null,!0,!1);let i=o.length;function f(n){const t=c(n);return function(n){n<36||(o[n]=i,i=n)}(n),t}function d(n){i===o.length&&o.push(o.length+1);const t=i;return i=o[t],o[t]=n,t}let _=new("undefined"==typeof TextDecoder?(0,n.require)("util").TextDecoder:TextDecoder)("utf-8",{ignoreBOM:!0,fatal:!0});_.decode();let a=null;function l(){return null!==a&&a.buffer===u.f.buffer||(a=new Uint8Array(u.f.buffer)),a}function b(n,t){return _.decode(l().subarray(n,n+t))}function w(n,t,e){u.h(n,t,d(e))}let s=0;let g=new("undefined"==typeof TextEncoder?(0,n.require)("util").TextEncoder:TextEncoder)("utf-8");const h="function"==typeof g.encodeInto?function(n,t){return g.encodeInto(n,t)}:function(n,t){const e=g.encode(n);return t.set(e),{read:n.length,written:e.length}};let p=null;function v(){return null!==p&&p.buffer===u.f.buffer||(p=new Int32Array(u.f.buffer)),p}function y(){u.g()}function m(n){return function(){try{return n.apply(this,arguments)}catch(n){u.a(d(n))}}}function x(n){return null==n}const E=function(n){c(n).preventDefault()},C=function(n){f(n)},j=function(n,t,e){c(n).textContent=0===t?void 0:b(t,e)},T=function(n){c(n).remove()},A=function(n){const t=f(n).original;if(1==t.cnt--)return t.a=0,!0;return!1},D=function(n){var t=function(n,t,e){if(void 0===e){const e=g.encode(n),r=t(e.length);return l().subarray(r,r+e.length).set(e),s=e.length,r}let r=n.length,u=t(r);const o=l();let c=0;for(;c<r;c++){const t=n.charCodeAt(c);if(t>127)break;o[u+c]=t}if(c!==r){0!==c&&(n=n.slice(c)),u=e(u,r,r=c+3*n.length);const t=l().subarray(u+c,u+r);c+=h(n,t).written}return s=c,u}(get_state(),u.c,u.d),e=s;v()[n/4+1]=e,v()[n/4+0]=t},k=function(n){return d(c(n))},O=function(n){return c(n)instanceof Window},P=function(n){var t=c(n).document;return x(t)?0:d(t)},q=function(n){var t=c(n).body;return x(t)?0:d(t)},F=function(n){var t=c(n).firstElementChild;return x(t)?0:d(t)},L=function(n){var t=c(n).nextElementSibling;return x(t)?0:d(t)},N=function(){return Date.now()},S=m((function(n,t,e){return d(c(n).createElement(b(t,e)))})),B=m((function(n,t){return d(c(n).appendChild(c(t)))})),I=m((function(n,t,e,r,u){c(n).setAttribute(b(t,e),b(r,u))})),z=function(n){var t=c(n).firstChild;return x(t)?0:d(t)},J=function(n){var t=c(n).nextSibling;return x(t)?0:d(t)},W=m((function(n,t,e,r){c(n).addEventListener(b(t,e),c(r))})),M=m((function(n,t){return d(c(n).cloneNode(0!==t))})),U=m((function(n,t,e,r){c(n).removeEventListener(b(t,e),c(r))})),G=function(n,t,e){c(n).className=b(t,e)},H=m((function(){return d(self.self)})),K=m((function(){return d(window.window)})),Q=m((function(){return d(globalThis.globalThis)})),R=m((function(){return d(r.global)})),V=function(n){return void 0===c(n)},X=function(n,t){return d(new Function(b(n,t)))},Y=m((function(n,t){return d(c(n).call(c(t)))})),Z=function(n,t){throw new Error(b(n,t))},$=function(n,t,e){return d(function(n,t,e,r){const o={a:n,b:t,cnt:1,dtor:e},c=(...n)=>{o.cnt++;try{return r(o.a,o.b,...n)}finally{0==--o.cnt&&(u.b.get(o.dtor)(o.a,o.b),o.a=0)}};return c.original=o,c}(n,t,22,w))}}).call(this,e(4)(n),e(5))},function(n,t,e){"use strict";var r=e.w[n.i];n.exports=r;e(2);r.i()},function(n,t){n.exports=function(n){if(!n.webpackPolyfill){var t=Object.create(n);t.children||(t.children=[]),Object.defineProperty(t,"loaded",{enumerable:!0,get:function(){return t.l}}),Object.defineProperty(t,"id",{enumerable:!0,get:function(){return t.i}}),Object.defineProperty(t,"exports",{enumerable:!0}),t.webpackPolyfill=1}return t}},function(n,t){var e;e=function(){return this}();try{e=e||new Function("return this")()}catch(n){"object"==typeof window&&(e=window)}n.exports=e}]]);
