--- frameworks/non-keyed/yarte-wasm-app/bundled-dist/index.js
@@ -1,1 +1,0 @@
-!function(e){function n(n){for(var t,o,u=n[0],i=n[1],c=0,_=[];c<u.length;c++)o=u[c],Object.prototype.hasOwnProperty.call(r,o)&&r[o]&&_.push(r[o][0]),r[o]=0;for(t in i)Object.prototype.hasOwnProperty.call(i,t)&&(e[t]=i[t]);for(a&&a(n);_.length;)_.shift()()}var t={},r={0:0};var o={};var u={3:function(){return{"./index_bg.js":{__wbg_preventDefault_7670dc6ff59bc226:function(e){return t[2].exports.r(e)},__wbindgen_object_drop_ref:function(e){return t[2].exports.D(e)},__wbg_settextContent_d2083d38d155212a:function(e,n,r){return t[2].exports.x(e,n,r)},__wbg_remove_753943fab80b89c7:function(e){return t[2].exports.t(e)},__wbindgen_cb_drop:function(e){return t[2].exports.z(e)},__wbg_getstate_7e06669f5fd16126:function(e){return t[2].exports.j(e)},__wbindgen_object_clone_ref:function(e){return t[2].exports.C(e)},__wbg_instanceof_Window_e8f84259147dce74:function(e){return t[2].exports.m(e)},__wbg_document_d3b6d86af1c5d199:function(e){return t[2].exports.g(e)},__wbg_body_61c142aa6eae691f:function(e){return t[2].exports.c(e)},__wbg_firstElementChild_f7ee091915e413c4:function(e){return t[2].exports.i(e)},__wbg_nextElementSibling_aab6b4e0e875e3b9:function(e){return t[2].exports.o(e)},__wbg_now_ad54c8b6865a8103:function(){return t[2].exports.q()},__wbg_createElement_d00b8e24838e36e1:function(e,n,r){return t[2].exports.f(e,n,r)},__wbg_appendChild_8658f795c44d1316:function(e,n){return t[2].exports.b(e,n)},__wbg_setAttribute_156f15ecfed9f628:function(e,n,r,o,u){return t[2].exports.v(e,n,r,o,u)},__wbg_firstChild_6278272abfd8998c:function(e){return t[2].exports.h(e)},__wbg_nextSibling_051d5a3751133ac7:function(e){return t[2].exports.p(e)},__wbg_addEventListener_116c561435e7160d:function(e,n,r,o){return t[2].exports.a(e,n,r,o)},__wbg_cloneNode_c0331876693c38ac:function(e,n){return t[2].exports.e(e,n)},__wbg_removeEventListener_d9ceb7fdf4ca5166:function(e,n,r,o){return t[2].exports.s(e,n,r,o)},__wbg_setclassName_5c90ce56028589a5:function(e,n,r){return t[2].exports.w(e,n,r)},__wbg_self_179e8c2a5a4c73a3:function(){return t[2].exports.u()},__wbg_window_492cfe63a6e41dfa:function(){return t[2].exports.y()},__wbg_globalThis_8ebfea75c2dd63ee:function(){return t[2].exports.k()},__wbg_global_62ea2619f58bf94d:function(){return t[2].exports.l()},__wbindgen_is_undefined:function(e){return t[2].exports.B(e)},__wbg_newnoargs_e2fdfe2af14a2323:function(e,n){return t[2].exports.n(e,n)},__wbg_call_e9f0ce4da840ab94:function(e,n){return t[2].exports.d(e,n)},__wbindgen_throw:function(e,n){return t[2].exports.E(e,n)},__wbindgen_closure_wrapper161:function(e,n,r){return t[2].exports.A(e,n,r)}}}}};function i(n){if(t[n])return t[n].exports;var r=t[n]={i:n,l:!1,exports:{}};return e[n].call(r.exports,r,r.exports,i),r.l=!0,r.exports}i.e=function(e){var n=[],t=r[e];if(0!==t)if(t)n.push(t[2]);else{var c=new Promise((function(n,o){t=r[e]=[n,o]}));n.push(t[2]=c);var _,f=document.createElement("script");f.charset="utf-8",f.timeout=120,i.nc&&f.setAttribute("nonce",i.nc),f.src=function(e){return i.p+""+({}[e]||e)+".js"}(e);var a=new Error;_=function(n){f.onerror=f.onload=null,clearTimeout(s);var t=r[e];if(0!==t){if(t){var o=n&&("load"===n.type?"missing":n.type),u=n&&n.target&&n.target.src;a.message="Loading chunk "+e+" failed.\n("+o+": "+u+")",a.name="ChunkLoadError",a.type=o,a.request=u,t[1](a)}r[e]=void 0}};var s=setTimeout((function(){_({type:"timeout",target:f})}),12e4);f.onerror=f.onload=_,document.head.appendChild(f)}return({1:[3]}[e]||[]).forEach((function(e){var t=o[e];if(t)n.push(t);else{var r,c=u[e](),_=fetch(i.p+""+{3:"85dcb0c98b21899a715e"}[e]+".module.wasm");if(c instanceof Promise&&"function"==typeof WebAssembly.compileStreaming)r=Promise.all([WebAssembly.compileStreaming(_),c]).then((function(e){return WebAssembly.instantiate(e[0],e[1])}));else if("function"==typeof WebAssembly.instantiateStreaming)r=WebAssembly.instantiateStreaming(_,c);else{r=_.then((function(e){return e.arrayBuffer()})).then((function(e){return WebAssembly.instantiate(e,c)}))}n.push(o[e]=r.then((function(n){return i.w[e]=(n.instance||n).exports})))}})),Promise.all(n)},i.m=e,i.c=t,i.d=function(e,n,t){i.o(e,n)||Object.defineProperty(e,n,{enumerable:!0,get:t})},i.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},i.t=function(e,n){if(1&n&&(e=i(e)),8&n)return e;if(4&n&&"object"==typeof e&&e&&e.__esModule)return e;var t=Object.create(null);if(i.r(t),Object.defineProperty(t,"default",{enumerable:!0,value:e}),2&n&&"string"!=typeof e)for(var r in e)i.d(t,r,function(n){return e[n]}.bind(null,r));return t},i.n=function(e){var n=e&&e.__esModule?function(){return e.default}:function(){return e};return i.d(n,"a",n),n},i.o=function(e,n){return Object.prototype.hasOwnProperty.call(e,n)},i.p="bundled-dist/",i.oe=function(e){throw console.error(e),e},i.w={};var c=window.webpackJsonp=window.webpackJsonp||[],_=c.push.bind(c);c.push=n,c=c.slice();for(var f=0;f<c.length;f++)n(c[f]);var a=_;i(i.s=0)}([function(e,n,t){t.e(1).then(t.bind(null,1)).catch(console.error)}]);
